switch to the newest libgdx ui

+++ c2d-tests/src/info/u250/c2d/tests/Nehe07_TextureFilters_Lighting.java
@@ -335,7 +335,7 @@ public class Nehe07_TextureFilters_Lighting extends Engine {
 							return super.fling(velocityX, velocityY);
 						};
 						@Override
-						public boolean tap(int x, int y, int count) {
+						public boolean tap(float x, float y, int count) {
 							i++;
 							if(i==TextureFilter.values().length)i=0;
