+++ c2d-tests/src/info/u250/c2d/tests/Nehe16_Fog.java
@@ -347,7 +347,7 @@ public class Nehe16_Fog extends Engine {
 							return super.fling(velocityX, velocityY);
 						};
 						@Override
-						public boolean tap(int x, int y, int count) {
+						public boolean tap(float x, float y, int count) {
 							i++;
 							if(i>=fogMode.length){
 								i = 0;
