+++ c2d-tests/src/info/u250/c2d/tests/Nehe18_Quadrics.java
@@ -203,7 +203,7 @@ public class Nehe18_Quadrics extends Engine {
 							return super.fling(velocityX, velocityY);
 						};
 						@Override
-						public boolean tap(int x, int y, int count) {
+						public boolean tap(float x, float y, int count) {
 							i++;
 							if(i==6){
 								i = 0;
